vterm-minibuffer: mv file, fixed term opening in wrong dir
I file wasn't being loaded when it was placed in the lisp/
directory. Additionally I found some advice on the internet suggesting
that placing files in the lisp/ directory probably wasn't the best, as
not all files in there are meant to be loaded (some are smartly loaded
by plugins depending on the architecture). So I moved the file to the
root directory and explicitly included it.

I also fixed a bug where if you called vterm-minibuffer it would open
a shell with the wrong PWD. Although vterm-minibuffer was determining
the base directory correctly using the version control dir, it wasn't
changing to that directory. This caused the vterm shell to open in the
wrong directory.